Only allow users in admin group to delete composes
Currently only ALLOWED_GROUPS can create and delete composes, however we
may want to grant the creating and deleting permissions to different
groups of users, so add ADMIN_GROUPS to only allow users in such groups
to delete composes.

Since we don't manage groups in odcs db, to grant a user with delete
permission, the user must be added to a (or some) particular group(s)
in the backend auth system first, and then add the group(s) to
ADMIN_GROUPS in config.