Add fedmsg support.
This publishes to fedmsg, if configured.

An important limitation here is that since we currently have only on
`TestConfiguration` for the project, it is not possible (or I don't see
a way) to have the test suite run *both* the rhmsg tests *and* the
fedmsg tests.  You can change the configuration for the tests and run
one, or the other.