Check requested packages are present in generated pungi compose
There have been a bug in pungi and it did not generate compose
with all the requested RPMs, but still returned zero error code.
We should prevent that by cheking the compose/metadata/rpms.json
in resulting compose to find out if all the packages we requested
are there and if not, fail the compose.

FIXES: #106