Prevent from using relative path for compose target_dir
When a relative path is set for target_dir, within the python process,
it's the relative path to where python process run from, but we will run
pungi command with the workdir set to a temporary, and then the
target_dir passed to pungi is actually a relative path to that temp dir,
that's not what we want and will cause problem.

For example, when we set target_dir as './composes' and start our
frontend or backend from /home/xxx/odcs, the target_dir is known as
'/home/xxx/odcs/composes', but pungi will generate composes under a
temporary dir, e.g. /tmp/tmpa1ncjqto/composes, since we have:

    td = tempfile.mkdtemp()
    ...
    odcs.server.utils.execute_cmd(pungi_cmd, cwd=td)

in pungi.py.