Run test against python3 rather than python3.5
Koji is required for running tests, but is not available on pypi, so we
need to enable sitepackages for tox, the problem is when we run on a
system (e.g. Fedora 26) that has python3.6 as default python3, we need
to copy koji package from 3.6's sitepackages to 3.5's sitepackage,
and python-rpm which is required by koji has shared libararies which
can't be easy copied (though we can hack it).

By change testenv py35 to py3, our tests will be ran against the default
python3 on system.