Don't pass pkgset_koji_tag=None to Pungi
For a modular compose, pkgset_koji_tag was set to None with the comment
that it had to be set to something, but current Pungi expects that
pkgset_koji_tag is unset, or a tag, or a list of tags, or the special
value 'not-used', which is like unset but with a warning.

See also https://pagure.io/pungi/pull-request/845 which is needed to
fully make Pungi do what it was meant to do. None doens't work in any
case.