Fix python3 issue caused by comparing `None` and `int` compose.koji_event.
Comparing `None` and `int` is not possible in python3, but works in python2.
The compose.koji_event is unset for modular composes, because the modular
Koji tag does not change once the module is built and therefore it is
not needed to track the koji_even on which the compose was generated.