Remove old cached data from Koji tag cache directory.
Each directory in the Koji tag cache directory is unique to particular
compose input data. In case when for example compose is generated from
particular Koji tag just once and never again, the cached data would
stay in NFS storage forever.

This is not good, because the NFS storage is not big and generally
we should cleanup the unused data.

In this commit, the cached data which has not been used for more than
30 days are removed. This is configurable.