Add "pungi_compose" source type.
When using this source_type, the "source" should contain URL to variant
repository of external compose generated by the Pungi. For example
https://kojipkgs.fedoraproject.org/compose/rawhide/latest-Fedora-Rawhide/compose/Server/.

The generated compose will contain the same set of RPMs as the given
external compose variant. The packages will be taken from the configured
Koji instance.

The use-case here is to be able to mirror existing releng compose in
a reproducible way. While typical releng composes are removed after few
days/weeks, the ODCS keeps the list of builds/packages in the database
and therefore can regenerate the same compose later when needed.