Add support for hybrid composes - composes with both modular and normal RPMs.
This adds `modular_koji_tags` list to API which can list the Koji tags
in which the modular Content Generator Koji builds created by MBS
are tagged. Such builds are included in a compose together with RPMs
from classic Koji tag defined in compose `source`. The "hybrid" Pungi
compose gather method is used to achieve that internally.

This adds `module_defaults_url` string to API which can be set to git
repository with module defaults which are added to the resulting
compose metadata or which are used when resolving RPMs for compose
when using Pungi hybrid compose gather method.