Allow creating composes with -devel modules.
The -devel modules do not exist in the MBS database, but they exist
only as CG build in Koji. Therefore the current code fails, because
it only queries MBS to get the modules.

This is also made worse by a fact that Pungi actually cannot handle
-devel CG module builds because of known issue. Therefore we cannot
simply pass the module NSVCs down to Koji, but we need to handle
them in a special way.

This commit detects the situation when -devel NSVC cannot be found
in the MBS and in this case it tries to find the non-devel version.
This is needed to be able to resolve uncomplete NSVC (for example
just name:stream) to full NSVC.

The -devel modules are then passed down to Pungi using the
`include_devel_modules` option, while non-devel modules are stored
in variant.xml as usually.