Include all modules in modular Koji tag in compose.
There was Pungi change when generating compose with modules from modular
Koji tag. To include all the modules which are in a tag, we now need
to use wildcard sign.

It is also possible to cherry-pick some modules from the modular
Koji tag, but our API currently does not support that.