Save merged repos in the subdirectory according to content-set.
This fixes an issue when first merged repository was overwriten by
the second one in case we were generating two merged repos in single
compose.

Previously, the merged repo was generated directly in
`$results_repo_dir/`. In this commit, it is generated in the
`$results_repo_dir/$content_set_name/`.