Store raw_config composes in the same directory layout as normal ones.
Before this commit, the directory in which the compose was generated
was defined by Pungi. Pungi defines it according to name of a product
for which the compose is generated, the release number, compose type,
and lot of other things.

This used to work correctly for composes generated from ODCS templates,
because their name/release/type is predefined and always the same and
therefore the compose directory was also well known.

With raw_config composes, we no longer know the product's name, release,
or other things. The directory name is therefore not predictable
which makes it imposible to point to a resulting compose or remove
it once it expires.

In this commit, the compose directory path is generated and created
in ODCS (not in Pungi anymore) and is passed to Pungi using the
--compose-dir command line argument. We generate the compose directory
also for raw_config composes and therefore they are stored in directory
with predictable name on filesystem.

This allows to remove these directories when compose expires.