Log Pungi stderr/stdout to logs in compose directory.
With raw_config, it can easily happen that the configuration file is
wrong and Pungi even cannot parse it. This is currently hard to debug,
because such error is not logged anywhere.

In this commit, the Pungi stdout/stderr is logged to log files in
the compose directory and therefore easy to access and check.