Add "odcs-mock-runroot" script to allow generating runroot tasks.
This is related to qcow2 image builds. In order to build such images,
Pungi needs to run "lorax" as root. To do that, we will use OpenSSH
runroot method and connect the runroot VM where such runroot task
will run in a Mock chroot.

The Pungi on ODCS backend will be configured to execute
`odcs-mock-runroot` script over SSH which will wrap the Mock calls.

This script is implemented in this commit.