Continue removing expired compose despite the error in `shutil.rmtree`.
The current code simply throws an exception on any error in `shutil.rmtree`
and stops removing expired composes completely. The error can for example
be "Permission denied" when some files in compose directory are owned
by root:root as a result of issue with runroot tasks execution.

In this commit, such errors are logged, but the rmtree execution continues
and removes as much composes/files as it can.