Remove the Mock chroot in case "init" or "install" phase fails.
So far we only removed the Mock chroot after the "run" phase, but in
case there is some issue before we even get to that phase during the
Mock "init" or "install" phases, the chroot data would remain on
filesystem forever.

In this commit, this is fixed by catching any issue in "init" and
"install" phases and removing the Mock chroot before re-raising.