Remove the Mock chroot also in case of exception in runroot command.
Before this commit, in the case of some error when running the
runroot command, the Mock chroot for this runroot command was
never removed. This is big issue, because these chroots could
consume all the free space on root fs.

In this commit, the Mock chroot is removed on any issue with the
runroot command execution.