Add new `include_done_modules` flag to generate compose with modules in `done` state.
Currently, only modules in the `ready` state can appear in a compose.
This is correct for most cases, but the ODCS is also used by gating
CI which needs to generate compose from just-built module in the `done`
state in order to test it.

New flag is added in this commit to allow this behaviour.

Additionally, if full NSVCs are specified as compose input, the module can be
in `done` state even without the `include_done_modules` flag. The reason is
that the requester asked for this NSVC directly, so ODCS should generate the
compose without checking the state.