odcs_test_deployment: Use py3 and list all the RPMs included in resulting compose.
With recent DNF changes, the modular RPMs were ignored by the DNF API we used.
The tests checking the RPMs in the composes containing modules therefore failed.

This commit fixes it by `hawkey.IGNORE_EXCLUDES` flag. This has been discussed
with DNF team and they advised to use this flag.