Update time_to_expire when compose is finished
When a compose transitions to "done" or "failed", update the expiration
time. Since we don't store the time to live for the compose, we have to
compute it based on start time and current expiration time.

This should prevent composes that took long time to finish from being
expired right away after finishing.

JIRA: FACTORY-4864