Set Celery's "server_hostname" only if it's requested in the config file.
In Fedora deployment, we don't want to use OpenSSL SNI feature which is
enabled by the "server_hostname". We therefore want to disable it if it's
not specifically requested in the ODCS configuration file.

In Fedora, we also use SSL auth, so this commit allows setting the SSL
certs using the CELERY_CONFIG dict.