Handle two repos with the same relative_url in the single content set.
Without this commit, when there are two repos in single content set
sharing the same relative_url, ODCS tries to create and lock single
file lock twice, because the locks are based on the relative_url. This
leads to deadlock.

It also calls mergerepo_c with the same repo twice which is waste
of resources.

This commit changes it, so we remove duplicated repos before calling
mergerepo_c.

It also adds 1800 seconds timeout to mergerepo_c to ensure it does not
block whole process forever in case of issues.