Add ignore_absent_pulp_repos flag
For `pulp` source, if a content set is specified but does not exist in
Pulp instance, an error is returned to the caller and no repofile will
be created.

When this flag is set, any content set absent in Pulp is ignored. This
is logged on the server, but the user will not see this.

If all of them are missing, an empty repo file will be produced.

JIRA: COMPOSE-4008