Add `odcs-promote-compose` server/contrib script.
This script should be used in the future to promote the ODCS compose
to candidate compose which can later be released.

So far the script does the basic sanity checks and copies compose
to target directory. During copying, symlinks are replaced with
hardlinks.

The script might in the future do also other needed Compose metadata
transformations.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>