Limit Pungi error log parsing output to 2kB.
Before this commit, the ODCS would store unlimited amount
of data comming from Pungi log in `state_reason`. This is
a problem, because Pungi can print lot of errors for example
if it cannot find any of the requested RPMs on filesystem.

This commit limits the error message parsed from Pungi logs
to 2kB.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>