Use only latest repo from Pulp
If Pulp returns multiple repos for a given content set, and all of them
have a single product version each, and the product version follows the
same schema for all repos, we can use only the latest one.

This is based on the assumption that X.Y+1 contains all the content from
X.Y, so there's no need to merge all the repodata.

This should hopefully make pulp composes faster and avoid issues where
mergerepo_c is running out of memory.

Fixes: https://pagure.io/odcs/issue/357
JIRA: COMPOSE-4231