doc: clarify API behavior with ODCS querying MBS
Remove the passive voice so that readers will understand which system
queries MBS (ODCS does it).

Use the word "build" instead of "version" to make it clear that ODCS is
searching MBS for any developer's latest build of a given module:stream.

Document that ODCS will only consider the latest modules from MBS that
are in the "ready" state unless the user sets the include_done_modules
flag.