Support regenerating raw_config composes using the stored pungi_config_dump.
This commit also changes the input to Pungi from classic pungi.conf
to pungi.json for all composes to keep the code consistent.

This is needed, because there is no reliable way to convert
pungi_config_dump from JSON to PyConfigParser format. Using JSON
instead for all Pungi configuration files passes to Pungi seems
to be a better option.