Do not use latest MBS module if compose.source is not defined.
The compose.source can be set to an emptry string in case we
want to generate compose only from scratch module builds.

The `MBS.get_latest_modules` does not count with that and in case
the nsvc is an empty string, it simply does not send `nsvc`
in the request to MBS which means all the modules are returned and
MBS class chooses the latest one and adds it to "source" later.

This commit handles this situation and returns early with an empty
list in this case.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>