Fix query repo by id
Pulp responses with status code 404 if no repository is found with the
given id. The _rest_get raises HTTPError when that happens, but the
original code didn't catch and handle the error correctly.

This change fixes the wrong behavior. Since the 404 response is the only
one to be handled, any other potential errors responded from Pulp server
will be raised continuously.

JIRA: CLOUDBLD-3796

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>