Allow setting builds and modules in PATCH to support compose respin.
This commit allows to override `modules` and `builds` when renewing
the compose. This should be used for compose respins when there is
a respin of compose with extra builds or modules to hotfix some issue
needed.

This commit also automatically sents the `Compose.respin_of` according
to `Compose.pungi_compose_id` of the original compose.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>