Add filters for base module stream version
MBS exposes the stream as so called "stream version", which allows
comparisons to be done on it. We can filter on it being less than or
equal to given value or greater than or equal.

This basically means a developer can ask for a compose that contains the
latest module compatible with X.Y.Z, and they will get a correct
response even if there is nothing built specifically for that version,
but a build exists for X.(Y-1).Z or some other previous version.

Request for <=X.Y.Z should also be accompanied by a filter for >= X.0.0,
but responsibility for adding that is on the caller.

JIRA: RHELCMP-4006