utils: Wait a short time for timer event to be set
The Timer implementation first runs the code of the function (which
kills the subprocess) and then sets the finished event to True.

This opens a possibility of a race condition, if the timer is checked in
the period after killing the subprocess but before event is set.

To make this less likely, we can wait for the even for a short period of
time. Ideally we would want the timer to be atomic, but that would
require reimplementing it.

Signed-off-by: Lubomír Sedlář <lsedlar@redhat.com>