mbs: Use base module stream when resolving dependencies
When a module for a particular release is requested, its dependencies
should not be pulled from newer release.

This is achieved by adding a filter for base_module_br_stream_version,
which has to be at most the same as for the requested module. We also
add a requirement that the same major version has to be used.

JIRA: RHELCMP-4592